pyproject.toml
[project]
name = "container-opti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "ruff>=0.8.0",
    "uuid>=1.30",
]
